Update dotfiles last
Homebrew installation can fail if the git configuration is changed first since I use git+ssh by default. That only works if I've generated an SSH key and put it on Github, which I don't actually usually do before bootstrapping. Reordering in this way removes the need for that assumption.